fixes and modernizes pubsub api test
