+++ pubsub/api/composer.json
@@ -6,28 +6,25 @@
     },
     "require-dev": {
         "phpunit/phpunit": "~4.8",
-        "google/cloud-tools": "^0.6"
+        "google/cloud-tools": "^0.8.5"
     },
     "autoload": {
-        "psr-4": {
-            "Google\\Cloud\\Samples\\PubSub\\": "src"
-        },
         "files": [
-            "src/functions/create_subscription.php",
-            "src/functions/create_topic.php",
-            "src/functions/create_push_subscription.php",
-            "src/functions/delete_subscription.php",
-            "src/functions/delete_topic.php",
-            "src/functions/get_subscription_policy.php",
-            "src/functions/get_topic_policy.php",
-            "src/functions/list_subscriptions.php",
-            "src/functions/list_topics.php",
-            "src/functions/publish_message.php",
-            "src/functions/pull_messages.php",
-            "src/functions/set_subscription_policy.php",
-            "src/functions/set_topic_policy.php",
-            "src/functions/test_subscription_permissions.php",
-            "src/functions/test_topic_permissions.php"
+            "src/create_subscription.php",
+            "src/create_topic.php",
+            "src/create_push_subscription.php",
+            "src/delete_subscription.php",
+            "src/delete_topic.php",
+            "src/get_subscription_policy.php",
+            "src/get_topic_policy.php",
+            "src/list_subscriptions.php",
+            "src/list_topics.php",
+            "src/publish_message.php",
+            "src/pull_messages.php",
+            "src/set_subscription_policy.php",
+            "src/set_topic_policy.php",
+            "src/test_subscription_permissions.php",
+            "src/test_topic_permissions.php"
         ]
     }
 }
