--- pubsub/api/src/ProjectIdTrait.php
@@ -1,36 +1,0 @@
-<?php
-/**
- * Copyright 2016 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace Google\Cloud\Samples\PubSub;
-
-/**
- * Trait for determining the current project ID using "gcloud"
- */
-trait ProjectIdTrait
-{
-    private function getProjectIdFromGcloud()
-    {
-        exec("gcloud config list --format 'value(core.project)' 2>/dev/null", $output, $return_var);
-
-        if (0 === $return_var) {
-            return array_pop($output);
-        }
-
-        throw new \Exception('Could not derive a project ID from gcloud. ' .
-            'You must supply a project ID using --project');
-    }
-}
