+++ pubsub/api/src/create_push_subscription.php
@@ -44,5 +44,7 @@ function create_push_subscription($projectId, $topicName, $subscriptionName, $en
     $subscription->create([
         'endpoint' => $endpoint
     ]);
+
+    printf('Subscription created: %s' . PHP_EOL, $subscription->name());
 }
 # [END pubsub_create_push_subscription]
