+++ pubsub/api/src/delete_topic.php
