+++ pubsub/api/src/list_subscriptions.php
