+++ pubsub/api/src/list_topics.php
